requirements.txt
aiohttp==3.8.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.13.0
atpublic==3.1.1
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
cryptography==38.0.4
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
dpath==2.1.3
dulwich==0.20.50
dvc==2.38.1
dvc-data==0.28.4
dvc-http==2.27.2
dvc-objects==0.14.0
dvc-render==0.0.15
dvc-task==0.1.8
dvclive==1.3.0
filelock==3.9.0
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.3.3
fsspec==2022.11.0
funcy==1.17
future==0.18.2
gitdb==4.0.10
GitPython==3.1.30
grandalf==0.6
hydra-core==1.3.1
idna==3.4
iterative-telemetry==0.0.6
kombu==5.2.4
multidict==6.0.4
nanotime==0.5.2
networkx==2.8.8
numpy==1.24.1
omegaconf==2.3.0
packaging==22.0
pandas==1.5.2
pathspec==0.9.0
prompt-toolkit==3.0.36
psutil==5.9.4
pycparser==2.21
pydot==1.4.2
pygit2==1.11.1
Pygments==2.13.0
pygtrie==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
pywin32==305
PyYAML==6.0
requests==2.28.1
rich==13.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scmrepo==0.1.4
shortuuid==1.0.11
shtab==1.5.8
six==1.16.0
smmap==5.0.0
tabulate==0.9.0
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
yarl==1.8.2
zc.lockfile==2.0



-e .